backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
drf-spectacular==0.29.0
pillow==12.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
openai==2.15.0
requests==2.32.5
beautifulsoup4==4.14.3
google-generativeai==0.7.2
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0

